+++ packages/core/accessibility/accessibility-properties.ts
@@ -31,6 +31,30 @@ export const accessibilityEnabledProperty = new CssProperty<Style, boolean>({
 });
 accessibilityEnabledProperty.register(Style);
 
+export const iosAccessibilityAdjustsFontSizeProperty = new InheritedCssProperty<Style, boolean>({
+	defaultValue: false,
+	name: 'iosAccessibilityAdjustsFontSize',
+	cssName: 'ios-a11y-adjusts-font-size',
+	valueConverter: booleanConverter,
+});
+iosAccessibilityAdjustsFontSizeProperty.register(Style);
+
+export const iosAccessibilityMinFontScaleProperty = new InheritedCssProperty<Style, number>({
+	defaultValue: 0,
+	name: 'iosAccessibilityMinFontScale',
+	cssName: 'ios-a11y-min-font-scale',
+	valueConverter: parseFloat,
+});
+iosAccessibilityMinFontScaleProperty.register(Style);
+
+export const iosAccessibilityMaxFontScaleProperty = new InheritedCssProperty<Style, number>({
+	defaultValue: 0,
+	name: 'iosAccessibilityMaxFontScale',
+	cssName: 'ios-a11y-max-font-scale',
+	valueConverter: parseFloat,
+});
+iosAccessibilityMaxFontScaleProperty.register(Style);
+
 export const accessibilityHiddenProperty = new (global.isIOS ? InheritedCssProperty : CssProperty)({
 	name: 'accessibilityHidden',
 	cssName: 'a11y-hidden',