+++ packages/core/file-system/file-system-access.android.ts
@@ -15,7 +15,7 @@ function getApplicationContext() {
 }
 
 function getOrSetHelper(path: string): org.nativescript.widgets.FileHelper {
-	return org.nativescript.widgets.FileHelper.fromString(applicationContext, path);
+	return org.nativescript.widgets.FileHelper.fromString(getApplicationContext(), path);
 }
 
 function isContentUri(path: string): boolean {
@@ -252,6 +252,42 @@ export class FileSystemAccess implements IFileSystemAccess {
 
 	public getCurrentAppPath(): string {
 		return this.getLogicalRootPath() + '/app';
+	}
+
+	public copy = this.copySync.bind(this);
+
+	public copySync(src: string, dest: string, onError?: (error: any) => any) {
+		try {
+			return org.nativescript.widgets.Async.File.copySync(src, dest, getApplicationContext());
+		} catch (error) {
+			if (onError) {
+				onError(exception);
+			}
+		}
+
+		return false;
+	}
+
+	public copyAsync(src: string, dest: string): Promise<boolean> {
+		return new Promise<boolean>((resolve, reject) => {
+			try {
+				org.nativescript.widgets.Async.File.copy(
+					src,
+					dest,
+					new org.nativescript.widgets.Async.CompleteCallback({
+						onComplete: (result: boolean) => {
+							resolve(result);
+						},
+						onError: (err) => {
+							reject(new Error(err));
+						},
+					}),
+					getApplicationContext()
+				);
+			} catch (ex) {
+				reject(ex);
+			}
+		});
 	}
 
 	public readBuffer = this.readBufferSync.bind(this);
@@ -730,6 +766,7 @@ export class FileSystemAccess29 extends FileSystemAccess {
 		if (isContentUri(path)) {
 			try {
 				const file = getOrSetHelper(path);
+
 				return {
 					path,
 					name: file.getName(),