+++ packages/core/ui/frame/fragment.transitions.android.ts
@@ -10,6 +10,7 @@ import lazy from '../../utils/lazy';
 import { Trace } from '../../trace';
 import { FadeTransition } from '../transition/fade-transition';
 import { SlideTransition } from '../transition/slide-transition';
+import { SharedTransition, SharedTransitionEventData, SharedTransitionAnimationType } from '../transition/shared-transition';
 
 interface TransitionListener {
 	new (entry: ExpandedEntry, transition: androidx.transition.Transition): ExpandedTransitionListener;
@@ -312,6 +313,27 @@ export function _reverseTransitions(previousEntry: ExpandedEntry, currentEntry: 
 	return transitionUsed;
 }
 
+function notifySharedTransition(id: number, eventName: string) {
+	const state = SharedTransition.getState(id);
+	if (!state) {
+		return;
+	}
+	SharedTransition.notifyEvent(eventName, {
+		id,
+		type: 'page',
+		action: state.activeType === SharedTransitionAnimationType.present ? 'present' : 'dismiss',
+	});
+	if (eventName === SharedTransition.finishedEvent) {
+		if (state.activeType === SharedTransitionAnimationType.present) {
+			SharedTransition.updateState(id, {
+				activeType: SharedTransitionAnimationType.dismiss,
+			});
+		} else {
+			SharedTransition.finishState(id);
+		}
+	}
+}
+
 // Transition listener can't be static because
 // android is cloning transitions and we can't expand them :(
 function getTransitionListener(entry: ExpandedEntry, transition: androidx.transition.Transition): ExpandedTransitionListener {
@@ -332,6 +354,9 @@ function getTransitionListener(entry: ExpandedEntry, transition: androidx.transi
 				if (Trace.isEnabled()) {
 					Trace.write(`START ${toShortString(transition)} transition for ${entry.fragmentTag}`, Trace.categories.Transition);
 				}
+				if (entry?.transition) {
+					notifySharedTransition(entry.transition?.id, SharedTransition.startedEvent);
+				}
 			}
 
 			onTransitionEnd(transition: androidx.transition.Transition): void {
@@ -341,6 +366,9 @@ function getTransitionListener(entry: ExpandedEntry, transition: androidx.transi
 					Trace.write(`END ${toShortString(transition)} transition for ${entry.fragmentTag} backEntry:${backEntry ? backEntry.fragmentTag : 'none'}`, Trace.categories.Transition);
 				}
 				transitionOrAnimationCompleted(entry, backEntry);
+				if (entry?.transition) {
+					notifySharedTransition(entry.transition?.id, SharedTransition.finishedEvent);
+				}
 				this.backEntry = null;
 			}