+++ packages/core/ui/image/image-common.ts
@@ -14,7 +14,7 @@ import { Trace } from '../../trace';
 @CSSType('Image')
 export abstract class ImageBase extends View implements ImageDefinition {
 	public imageSource: ImageSource;
-	public src: string | ImageSource;
+	public src: string | ImageSource | ImageAsset;
 	public isLoading: boolean;
 	public stretch: CoreTypes.ImageStretchType;
 	public loadMode: 'sync' | 'async';
@@ -134,7 +134,7 @@ export const imageSourceProperty = new Property<ImageBase, ImageSource>({
 });
 imageSourceProperty.register(ImageBase);
 
-export const srcProperty = new Property<ImageBase, any>({ name: 'src' });
+export const srcProperty = new Property<ImageBase, string | ImageSource | ImageAsset>({ name: 'src' });
 srcProperty.register(ImageBase);
 
 export const loadModeProperty = new Property<ImageBase, 'sync' | 'async'>({