+++ packages/core/ui/transition/modal-transition.ios.ts
@@ -1,7 +1,8 @@
 import type { View } from '../core/view';
+import { CORE_ANIMATION_DEFAULTS, getDurationWithDampingFromSpring } from '../../utils/common';
 import { isNumber } from '../../utils/types';
 import { Transition, SharedElementSettings, TransitionInteractiveState } from '.';
-import { SharedTransition, DEFAULT_DURATION } from './shared-transition';
+import { SharedTransition, SharedTransitionAnimationType } from './shared-transition';
 import { SharedTransitionHelper } from './shared-transition-helper';
 import { PanGestureEventData, GestureStateTypes } from '../gestures';
 
@@ -162,7 +163,26 @@ class ModalTransitionController extends NSObject implements UIViewControllerAnim
 	}
 
 	transitionDuration(transitionContext: UIViewControllerContextTransitioning): number {
-		return DEFAULT_DURATION;
+		const owner = this.owner.deref();
+		if (owner) {
+			const state = SharedTransition.getState(owner.id);
+			switch (state?.activeType) {
+				case SharedTransitionAnimationType.present:
+					if (isNumber(state?.pageEnd?.duration)) {
+						return state.pageEnd?.duration / 1000;
+					} else {
+						return getDurationWithDampingFromSpring(state.pageEnd?.spring).duration;
+					}
+
+				case SharedTransitionAnimationType.dismiss:
+					if (isNumber(state?.pageReturn?.duration)) {
+						return state.pageReturn?.duration / 1000;
+					} else {
+						return getDurationWithDampingFromSpring(state.pageReturn?.spring).duration;
+					}
+			}
+		}
+		return CORE_ANIMATION_DEFAULTS.duration;
 	}
 
 	animateTransition(transitionContext: UIViewControllerContextTransitioning): void {